app/requirements.txt
anyio==2.2.0
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==20.3.0
autopep8==1.6.0
Babel==2.9.0
backcall==0.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
click==7.1.2
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cycler==0.10.0
decorator==5.0.7
defusedxml==0.7.1
deprecation==2.1.0
dj-database-url==0.5.0
django-filter==21.1
django-heroku==0.0.0
django-mkdocs==0.1.2
django-parler-rest==2.1
django-parler==2.3
django-timezone-field==4.1.2
Django==3.2
djangorestframework==3.13.1
entrypoints==0.3
future==0.18.2
ghp-import==2.0.2
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.10.0
ipykernel==5.5.3
ipython-genutils==0.2.0
ipython==7.22.0
iso8601==0.1.14
jedi==0.18.0
Jinja2==2.11.3
joblib==1.1.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.9.1
jupyter-server==1.6.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
jupyterlab==3.0.14
kiwisolver==1.3.1
kombu==5.0.2
lxml==4.6.3
Markdown==3.3.6
MarkupSafe==1.1.1
matplotlib==3.4.1
mergedeep==1.3.4
mistune==0.8.4
mkdocs-include-markdown-plugin==3.2.3
mkdocs==1.2.3
nbclassic==0.2.7
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.2
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parler==1.0.1
parso==0.8.2
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pip==22.1.2
prometheus-client==0.10.1
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycodestyle==2.8.0
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
python-docx==0.8.10
python-dotenv==0.17.0
pytz==2021.1
pyyaml-env-tag==0.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
schedule==1.1.0
scikit-learn==1.0.1
scipy==1.6.2
seaborn==0.11.1
selenium==3.141.0
Send2Trash==1.5.0
serial==0.0.97
setuptools==60.2.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.3.1
sqlparse==0.4.1
statsmodels==0.13.1
terminado==0.9.4
testpath==0.4.4
threadpoolctl==3.0.0
toml==0.10.2
tomlkit==0.7.0
tornado==6.1
traitlets==5.0.5
urllib3==1.26.4
vine==5.0.0
voila==0.2.16
watchdog==2.1.6
wcwidth==0.2.5
webdriver-manager==3.4.2
webencodings==0.5.1
wheel==0.37.1
whitenoise==5.2.0
zipp==3.7.0
